refact[4.2]: separate app into own file
