+++ models/Blog.js
@@ -2,7 +2,11 @@ const mongoose = require('mongoose')
 
 const blogSchema = new mongoose.Schema({
   title: String,
-  author: String,
+  author: {
+    type: String,
+    minLength: 3,
+    required: true
+  },
   url: String,
   likes: Number
 })
