+++ utils/middleware.js
@@ -21,7 +21,7 @@ const errorHandler = (error, request, response, next) => {
     return response.status(400).json({ error: error.message })
   }
 
-  next(error)
+  return next(error)
 }
 
 module.exports = {
